Make analysis and deprecations configurable
This will need you to call it twice. Once for analysis and once for
deprecations.

We need this so we can run the two checks on different paths.

And we need that because experience shows we cannot run analysis on
contrib modules without exhausting memory.

But we would still like to know about deprecations in contrib modules.