Run on node instead of in docker
Although we documented the need to run
i.e. `shivammathur/setup-php@master` to setup a PHP environment we
actually didn't use it when running `drupal-check`.

Since we ran drupal-check in a Docker container the PHP version in the
Docker image was of course used.

The was a problem because you cannot specify your own PHP version and
PHP extensions then. And that might lead drupal-check detecting false
errors.

So we do a rewrite of most things and run it with via node.js instead
which runs in the same environment as the setup of PHP.

We also introduce a new parameter to set the report name (but we default
to `Analysis` and `Deprecations` so no need to set those).

The rewrite is API compatible with the previous version. So users of
this actions doesn't have to change anything.